improve util helper functions
