+++ input/otlp/traces_test.go
@@ -918,8 +918,8 @@ func TestSpanDataStream(t *testing.T) {
 		{
 			resourceDataStreamDataset:   otlp.DisallowedDataStreamRunes + randomString,
 			resourceDataStreamNamespace: otlp.DisallowedDataStreamRunes + randomString,
-			expectedDataStreamDataset:   "____________" + randomString[:maxLen],
-			expectedDataStreamNamespace: "____________" + randomString[:maxLen],
+			expectedDataStreamDataset:   revertedString(otlp.DisallowedDataStreamRunes) + randomString[:maxLen],
+			expectedDataStreamNamespace: revertedString(otlp.DisallowedDataStreamRunes) + randomString[:maxLen],
 		},
 	} {
 		for _, isTxn := range []bool{false, true} {
